@@ -1,13 +1,12 @@
 {
     "name": "devcoder-xyz/php-options-resolver",
-    "description": "Processes and validates option array",
+    "description": "This library provides a simple solution for processing and validating option arrays in PHP.",
     "type": "library",
     "license": "MIT",
     "authors": [
         {
             "name": " F. Michel.R",
-            "email": "dev@devcoder.xyz",
-            "homepage": "https://www.devcoder.xyz"
+            "email": "dev@devcoder.xyz"
         }
     ],
     "autoload": {
@@ -18,7 +17,7 @@
         }
     },
     "require": {
-        "php": ">=7.3"
+        "php": ">=7.4"
     },
     "require-dev": {
         "phpunit/phpunit": "~9.4"
